requirements.txt
discord
audioop-lts
python-dotenv